+++ test/language/array_types/cpp/AutoArrayTest.cpp
@@ -17,7 +17,7 @@ protected:
     {
         writer.writeVarUInt64(static_cast<uint64_t>(length));
         for (size_t i = 0; i < length; ++i)
-            writer.writeBits(i, 8);
+            writer.writeBits(static_cast<uint32_t>(i), 8);
     }
 
     void checkBitSizeOf(size_t numElements)
@@ -25,7 +25,7 @@ protected:
         zserio::UInt8Array uint8Array;
         uint8Array.reserve(numElements);
         for (size_t i = 0; i < numElements; ++i)
-            uint8Array.push_back(i);
+            uint8Array.push_back(static_cast<uint8_t>(i));
         AutoArray autoArray;
         autoArray.setUint8Array(uint8Array);
 
@@ -39,7 +39,7 @@ protected:
         zserio::UInt8Array uint8Array;
         uint8Array.reserve(numElements);
         for (size_t i = 0; i < numElements; ++i)
-            uint8Array.push_back(i);
+            uint8Array.push_back(static_cast<uint8_t>(i));
         AutoArray autoArray;
         autoArray.setUint8Array(uint8Array);
 
@@ -68,7 +68,7 @@ protected:
         zserio::UInt8Array uint8Array;
         uint8Array.reserve(numElements);
         for (size_t i = 0; i < numElements; ++i)
-            uint8Array.push_back(i);
+            uint8Array.push_back(static_cast<uint8_t>(i));
         AutoArray autoArray;
         autoArray.setUint8Array(uint8Array);
