+++ test/language/default_values/cpp/StructureDefaultValuesTest.cpp
@@ -51,7 +51,7 @@ TEST(StructureDefaultValuesTest, checkDefaultFloat32Value)
 TEST(StructureDefaultValuesTest, checkDefaultFloat64Value)
 {
     StructureDefaultValues structureDefaultValues;
-    float diff = 1.2345 - structureDefaultValues.getFloat64Value();
+    double diff = 1.2345 - structureDefaultValues.getFloat64Value();
     if (diff < 0.0)
         diff = -diff;
     ASSERT_TRUE(diff <= std::numeric_limits<double>::epsilon());
